# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Staffer",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login/start": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Login Start",
        "operationId": "login_start_api_v1_auth_login_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login/exchange": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Login Exchange",
        "operationId": "login_exchange_api_v1_auth_login_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginExchange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Refresh",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Refresh"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Logout",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/devices": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "Devices",
        "operationId": "devices_api_v1_devices_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/devices/{device_id}": {
      "delete": {
        "tags": [
          "authentication"
        ],
        "summary": "Revoke Device",
        "operationId": "revoke_device_api_v1_devices__device_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{device_id}/push-token": {
      "put": {
        "tags": [
          "authentication"
        ],
        "summary": "Register Push Token",
        "operationId": "register_push_token_api_v1_devices__device_id__push_token_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushToken"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginExchange": {
        "properties": {
          "flow_id": {
            "type": "string",
            "format": "uuid",
            "title": "Flow Id"
          },
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "verifier": {
            "type": "string",
            "maxLength": 128,
            "minLength": 43,
            "title": "Verifier"
          }
        },
        "type": "object",
        "required": [
          "flow_id",
          "device_id",
          "verifier"
        ],
        "title": "LoginExchange"
      },
      "LoginStart": {
        "properties": {
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "platform": {
            "type": "string",
            "enum": [
              "WINDOWS",
              "ANDROID",
              "windows",
              "android"
            ],
            "title": "Platform"
          },
          "device_label": {
            "type": "string",
            "maxLength": 255,
            "title": "Device Label",
            "default": "Personal device"
          },
          "challenge": {
            "type": "string",
            "maxLength": 43,
            "minLength": 43,
            "title": "Challenge"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "platform",
          "challenge"
        ],
        "title": "LoginStart"
      },
...
      "PushToken": {
        "properties": {
          "push_token": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 10,
            "title": "Push Token"
          },
          "platform": {
            "type": "string",
            "enum": [
              "ANDROID",
              "android"
            ],
            "title": "Platform"
          }
        },
        "type": "object",
        "required": [
          "push_token",
          "platform"
        ],
        "title": "PushToken"
      },
...
      "Refresh": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "maxLength": 256,
            "minLength": 32,
            "title": "Refresh Token"
          },
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          }
        },
        "type": "object",
        "required": [
          "refresh_token",
          "device_id"
        ],
        "title": "Refresh"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}